\name{beach}
\alias{beach}
\title{A simulated survey region comprising the beach on which the seal population 'seal.pop' is  
located.}
\description{
The region is based on an aerial photograph of a real seal population on a beach. 

\itemize{
\item{"length"}{Length of base of survey region.}
\item{"width"}{"Height" of survey region.}
\item{"created"}{Creation date}
}}
\usage{data(beach)}
\format{An object of class `population`}
\source{Aerial photograph}
\references{None}
\keyword{datasets}

