\name{dp.to.ddf}
\alias{dp.to.ddf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Conversion of double platform samples for analysis by MRDS }
\description{
  Convert a Wisp object of type 'sample.dp' into format acceptable to code{ddf}
}
\usage{
dp.to.ddf(the.sample)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{the.sample}{ Object of class 'sample.dp' to be converted for analysis by \code{ddf} }
}
\details{
  More sophisticated models for line transect data can be constructed using algorithms available in the
  \pkg{MRDS} library, written by Jeff Laake.  In particular, models incorporating multiple covariates can be 
  analyzed.  This routine transforms line transect samples generated by Wisp into a format compatible with 
  \pkg{MRDS}, more specifically the model-fitting component of MRDS, namely \code{ddf}.

}
\value{
  A data frame consisting of:
  \item{object }{Sequential number of objects detected during the survey}
  \item{observer }{Observer who made the detection (presently always 1)}
  \item{detected }{Whether the object was detected by the specified observer, presently always 1}
  \item{distance }{Distance of object from transect line}
  \item{size }{Group size of the detected cluster}
  \item{sex }{Gender of detected cluster (species exhibits sexual segregation, hence travel in single-sex groups)}
  \item{exposure }{Detectability factor of the detected group}  
}
\references{ Laake, J.L. and D.L. Borchers. 2004. Methods for incomplete detection at distance zero. In: Advanced Distance Sampling, eds. S.T. Buckland, D.R.Anderson, K.P. Burnham, J.L. Laake, D.L. Borchers, and L. Thomas. Oxford University Press.

Marques, F.F.C. and S.T. Buckland. 2004. Covariate models for the detection function. In: Advanced Distance Sampling, eds. S.T. Buckland, D.R.Anderson, K.P. Burnham, J.L. Laake, D.L. Borchers, and L. Thomas. Oxford University Press. 
}
\author{ Eric Rexstad, RUWPA \email{ericr@mcs.st-and.ac.uk} }

\seealso{ \code{\link{generate.population}}, \code{\link{generate.sample.dp}}, also consult documentation for
		the MRDS library, specifically \code{ddf}, \code{plot.ddf}, and \code{ddf.gof} }
\examples{
library(mrds)
data(harb.samp.dp)
test <- dp.to.ddf(harb.samp.dp)
double.obs <- ddf(dsmodel = ~cds(key = "hn"), mrmodel = ~glm(~distance), data = test, method = "trial", meta.data = list(width = 1))
summary(double.obs)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{}

