\name{generate.design.dp}
\alias{generate.design.dp}
\title{Survey Design of Double Platform Method }
\description{
        This function uses the data stored in the `pars.design.dp´ object to create a new design object
}
\usage{
        generate.design.dp(pars.des, seed = NULL)
}
\arguments{
\item{pars.des}{An parameter object returned by the function \code{setpars.design.dp}}
\item{seed}{the number passed to set.seed() to initialise random number generator}
}
\details{
        A survey design of the double platform method is generated as design object. \code{generate.design.dp} works with the object `pars.design.dp´ that was generated by the function \code{setpars.design.dp}. \code{setpars.design.dp} has to be run prior.}
\value{
\code{generate.design.dp} returns an object of class `design.dp´ which has the following elements:
  \item{region}{object of class `region´}
  \item{n.transects}{number of transect paths}
  \item{effective.n.transects}{the effective number of transect paths}
  \item{n.units}{number of survey units}
  \item{pos.x}{vector of x-positions of the survey units}
  \item{start.y}{vector of starting y-positions of the survey units}
  \item{end.y}{vector of ending y-positions of the survey units}
  \item{visual.range}{visibility of both observers}
  \item{percentage.on.effort}{percentage of the transect lines that is searched}
  \item{parents}{Details of WiSP objects passed to function}
  \item{created}{Creation date and time}
  \item{seed}{Equal to the argument 'seed' passed to the function} 
}
\seealso{
        \code{\link{setpars.design.dp}}, \code{\link{plot.design.dp}}
}
\examples{
dp.reg <- generate.region(x.length = 100, y.width = 50)
dp.dens <- generate.density(dp.reg)

#heterogeneous population
dp.poppars<-setpars.population(density.pop = dp.dens, number.groups = 1000, size.method = "poisson",
                              size.min = 1, size.max = 30, size.mean = 10, exposure.method = "beta",
                              exposure.min = 0, exposure.max = 1, exposure.mean = 0.4,
                              exposure.shape = 0.5,type.values=c("Male","Female"), type.prob=c(0.48,0.52))
dp.pop<-generate.population(dp.poppars)

dp.despars<-setpars.design.dp(dp.reg, n.transects=10, n.units=10, visual.range=2, percent.on.effort=1)
dp.des<-generate.design.dp(dp.despars, seed=3)
summary(dp.des)
}
\keyword{}
