\name{generate.population}
\alias{generate.population}
\title{Defining an Animal Population }
\description{
        This function uses the data stored in the `pars.population´ object to create a new population object.
}
\usage{
        generate.population(pars.population, seed = NULL)
}
\arguments{
\item{pars.population}{An object created by the function \code{setpars.population} defining the parameters of the population.}
\item{seed}{the number passed to set.seed() to initialise random number generator}
}
\details{
        An animal population is generated as population object.  \code{generate.population} works with the object `pars.population´ that was generated by the function \code{setpars.population}. \code{setpars.population} has to  be run prior.
}
\value{
\code{generate.population} returns an object of class `population´ which has the following elements:
  \item{region}{object of class `region´}
  \item{groupID}{vector of identification numbers for the animal groups. }
  \item{posx}{vector of x-positions of the animal groups.}
  \item{posy}{vector of y-positions of the animal groups.}
  \item{groupsize}{vector of sizes of the animal groups.}
  \item{ntypes}{number of possible type properties for the animal groups.}
  \item{types}{vector of type properties of the animal groups.}
  \item{minexposure}{lowest possible group exposure value.}
  \item{maxexposure}{highest possible group exposure value.}
  \item{exposure}{vector of exposure values of the animal groups.}
  \item{parents}{Details of WiSP objects passed to function}
  \item{created}{Creation date and time}
  \item{seed}{Equal to the argument 'seed' passed to the function}
}
\seealso{
        \code{\link{setpars.population}}, \code{\link{summary.population}} 
		\code{\link{plot.population}}
}
\examples{
reg <- generate.region(x.length = 50, y.width = 80)
     
# set up a linear density, add a hotspot and a stripe
dens <- generate.density(reg, nint.x = 50, nint.y = 40, southwest = 1, southeast = 10, northwest = 20)
dens<-add.hotspot(dens, 20,10, 100,10)
dens<-set.stripe(dens, 2,5, 50,70, value = 0, width = 10)
pop.pars<-setpars.population(density.pop = dens, number.groups = 100, size.method = "poisson",
                             size.min = 1, size.max = 5, size.mean = 1, exposure.method = "beta",
                             exposure.min = 2, exposure.max = 10, exposure.mean = 6, exposure.shape = 1)

pop <- generate.population(pop.pars)
summary(pop)
plot(pop)
}
\keyword{}

