\name{generate.sample.dp}
\alias{generate.sample.dp}
\title{Survey Sample of Double Platform Method }
\description{
        This function uses the data stored in the `pars.survey.dp´ object to create a new sample object.
}
\usage{
        generate.sample.dp(pars.survey.dp)
}
\arguments{
\item{pars.survey.dp}{Parameter object returned by the function \code{setpars.survey.dp}}
}
\details{
        A survey sample of the double platform method is generated as sample object. \code{generate.sample.dp} works with the object `pars.survey.dp´ that was generated by the function \code{setpars.survey.dp}. \code{setpars.survey.dp} has to be run prior.
}
\value{
\code{generate.sample.dp} returns an object of class `sample.dp´ which has the following elements:
  \item{population}{object of class `population´.}
  \item{design}{object of class `design.dp´.}
  \item{detected.observer1}{vector indicating which animal groups have been detected by observer 1. \code{NA} means that an animal group was outside all survey units and could not be detected. \code{0} indicates that an animal group was inside a survey unit however has not been detected. \code{1} means that an animal group was inside a survey unit and has been detected by the observer.}
  \item{detected.observer2}{vector indicating which animal groups have been detected by observer 2. \code{NA} means that an animal group was outside all survey units and could not be detected. \code{0} indicates that an animal group was inside a survey unit however has not been detected. \code{1} means that an animal group was inside a survey unit and has been detected by the observer.}
  \item{distance}{vector of perpendicular distances of animal groups inside the survey units from the respective transect paths.}
  \item{transect}{vector of transect path numbers of detectable animal groups.}
}
\seealso{
        \code{\link{setpars.survey.dp}},   \code{\link{summary.sample.dp}}
		\code{\link{plot.sample.dp}}
}
\examples{
dp.reg <- generate.region(x.length = 100, y.width = 50)
dp.dens <- generate.density(dp.reg)

#heterogeneous population
dp.poppars<-setpars.population(density.pop = dp.dens, number.groups = 1000, size.method = "poisson",
                               size.min = 1, size.max = 30, size.mean = 10, exposure.method = "beta",
                               exposure.min = 0, exposure.max = 1, exposure.mean = 0.4,
                               exposure.shape = 0.5, type.values=c("Male","Female"), type.prob=c(0.48,0.52))
dp.pop<-generate.population(dp.poppars)
dp.despars<-setpars.design.dp(dp.reg, n.transects=10, n.units=10, visual.range=2, percent.on.effort=1)
dp.des<-generate.design.dp(dp.despars, seed=3)
dp.survpars<-setpars.survey.dp(dp.pop, dp.des, theta.obs1=0.7, theta.obs2=0.4, theta.exp=0.2, theta.dist=-0.7)

dp.samp<-generate.sample.dp(dp.survpars)
summary(dp.samp)
plot(dp.samp)
}
\keyword{}

