\name{generate.sample.lt}
\alias{generate.sample.lt}
\title{Line Transect Method Survey Data Generation}
\description{
        This function uses the data stored in the `pars.survey.lt´ object to create a line transect sample object.
}
\usage{
        generate.sample.lt(pars.survey.lt, seed = NULL)
}
\arguments{
\item{pars.survey.lt}{object returned by the function \code{setpars.survey.lt} defining the survey parameters.}
\item{seed}{number passed to set.seed() to initialise random number generator}
}
\details{
        A line transect method sample object is generated by simulating a line transect survey of the population specified in \code{generate.sample.lt}. The function \code{setpars.survey.lt} must have been run to create a `pars.survey.lt' object, which contains the survey specifations.
}
\value{ 
\code{generate.sample.lt} returns an object of class `sample.lt´ which has the following elements:
  \item{population}{object of class `population´.}
  \item{design}{object of class `design.lt´.}
  \item{detected}{vector indicating which animal groups have been detected. \code{NA} means that an animal group was outside all survey units and could not be detected. \code{0} indicates that an animal group was inside a survey unit however has not been detected. \code{1} means that an animal group was inside a survey unit and has been detected.}
  \item{distance}{vector of perpendicular distances of animal groups inside the survey units from the respective transect paths.}
  \item{transect}{vector of transect path numbers of detectable animal groups.}
  \item{parents}{Details of WiSP objects passed to function}
  \item{created}{Creation date and time}
  \item{seed}{Equal to the argument 'seed' passed to the function}
}
\seealso{
        \code{\link{setpars.survey.lt}},   \code{\link{summary.sample.lt}}
		\code{\link{plot.sample.lt}}
}
\examples{
lt.reg <- generate.region(x.length = 100, y.width = 50)
lt.dens <- generate.density(lt.reg)

#heterogeneous population
lt.poppars<-setpars.population(density.pop = lt.dens, number.groups = 1000,
                              size.method = "poisson", size.min = 1, size.max = 30,
                              size.mean = 10, exposure.method = "beta", exposure.min = 0,
                              exposure.max = 1, exposure.mean = 0.4, exposure.shape = 0.5,
                              type.values=c("Male","Female"), type.prob=c(0.48,0.52))
lt.pop<-generate.population(lt.poppars)
lt.despars<-setpars.design.lt(lt.reg, n.transects=10, n.units=10, visual.range=4, percent.on.effort=1)
lt.des<-generate.design.lt(lt.despars, seed=3)
lt.survpars<-setpars.survey.lt(lt.pop, lt.des, disthalf.min=2, disthalf.max=4)

lt.samp<-generate.sample.lt(lt.survpars)
summary(lt.samp)
plot(lt.samp)
}
\keyword{}

