\name{generate.sample.pl}
\alias{generate.sample.pl}
\title{Survey Sample of Plot Method }
\description{
        This function uses the data stored in the `pars.survey.pl´ object to create a plot sampling sample object.
}
\usage{
        generate.sample.pl(pop, des, seed = NULL)
}
\arguments{
\item{pop}{object of class `population´, returned by the function \code{generate.population}}
\item{des}{object of class `design.pl´ returned by the function \code{generate.design.pl}}
\item{seed}{number passed to set.seed() to initialise random number generator}
}
\details{
        A plot sampling method sample object is generated by simulating a plot sampling survey of the population `pop', using the design specified in `des'.
}
\value{
\code{generate.sample.pl} returns an object of class `sample.pl´ which has the following elements:
  \item{population}{object of class `population´.}
  \item{design}{object of class `design.pl´.}
  \item{detected}{vector indicating which animal groups have been detected. \code{0} indicates that an animal group has not been detected, \code{1} means that an animal group has been detected. }
  \item{unit}{vector indicating survey units (plots) in which each animal groups was detected.}
  \item{parents}{Details of WiSP objects passed to function}
  \item{created}{Creation date and time}
  \item{seed}{Equal to the argument 'seed' passed to the function}
}
\seealso{
        \code{\link{generate.design.pl}},   \code{\link{summary.sample.pl}}
		\code{\link{plot.sample.pl}},       \code{\link{set.seed}}
}
\examples{
pl.reg <- generate.region(x.length = 100, y.width = 50)
pl.dens <- generate.density(pl.reg)

pl.poppars<-setpars.population(density.pop=pl.dens, number.groups = 100, size.method = "poisson",
                               size.min = 1, size.max = 5, size.mean = 1, exposure.method = "beta",
                               exposure.min = 2, exposure.max = 10, exposure.mean = 6,
                               exposure.shape = 1)
pl.pop <- generate.population(pl.poppars, seed=456)

pl.despars<-setpars.design.pl(pl.reg, n.interval.x = 10, n.interval.y = 20,
                              method = "random",  area.covered = 0.2)
pl.des <- generate.design.pl(pl.despars, seed=789)

pl.samp<-generate.sample.pl(pl.pop, pl.des, seed=101112)
summary(pl.samp)
plot(pl.samp)
}
\keyword{}

