\name{generate.sample.pt}
\alias{generate.sample.pt}
\title{Point Transect Method Survey Data Generation}
\description{
     This function uses the data stored in the `pars.survey.pt´ object to create a point transect sample object.}

\usage{
     generate.sample.pt(pars.survey.pt, seed = NULL)}
\arguments{
\item{pars.survey.pt}{object returned by the function \code{setpars.survey.pt} defining the survey parameters.}
\item{seed}{number passed to set.seed() to initialise random number generator}
}
\details{
     A point transect method sample object is generated by simulating a point transect survey of the population specified in  
\code{generate.sample.pt}. The function \code{setpars.survey.pt} must have been run to create a `pars.survey.pt' object, which contains the survey  
specifations.}
\value{
     \code{generate.sample.pt} returns an object of class `sample.pt´ which has the following elements:

\item{population}{object of class `population´.}

\item{design}{object of class `design.pt´.}

\item{detected}{vector indicating which animal groups have been detected. \code{NA} means that an animal group was outside all survey units and could  
not be detected. \code{0} indicates that an animal group was inside a survey unit however has not been detected. \code{1} means that an animal group  
was inside a survey unit and has been detected.}

\item{distance}{vector of perpendicular distances of animal groups inside the survey
          units from the respective transect paths.}

\item{transect}{vector of transect path numbers of detectable animal groups.}
}
\seealso{
     \code{\link{setpars.survey.pt}}, \code{\link{summary.sample.pt}}, \code{\link{plot.sample.pt}}
}
\examples{
pt.reg <- generate.region(x.length = 100, y.width = 50)
pt.dens <- generate.density(pt.reg)

#heterogeneous population
pt.poppars<-setpars.population(density.pop = pt.dens, number.groups = 1000, size.method = "poisson",
                               size.min = 1, size.max = 30, size.mean = 10, exposure.method = "beta",
                               exposure.min = 0, exposure.max = 1, exposure.mean = 0.4,
                               exposure.shape = 0.5, type.values=c("Male","Female"),
                               type.prob=c(0.48,0.52))
pt.pop<-generate.population(pt.poppars)

pt.despars<-setpars.design.pt(pt.reg, n.transects=8, n.units=32, visual.range=3.5)
pt.des<-generate.design.pt(pt.despars)

pt.survpars<-setpars.survey.pt(pt.pop, pt.des, disthalf.min=2, disthalf.max=4)
pt.samp<-generate.sample.pt(pt.survpars)
summary(pt.samp)
plot(pt.samp)
}
\keyword{}

