\name{generate.sample.rm}
\alias{generate.sample.rm}
\title{Survey Sample of Removal Methods}
\description{
        This function uses the data stored in the `pars.survey.rm´ object to create a removal method sample object.
}
\usage{
        generate.sample.rm(pars.survey.rm, seed = NULL)
}
\arguments{
\item{pars.survey.rm}{object returned by the function \code{setpars.survey.rm} defining the survey parameters.}
\item{seed}{the number passed to set.seed() to initialise random number generator}
}
\details{
        A removal sample object is generated by simulating a removal survey of the population specified in \code{generate.sample.rm}. The function \code{setpars.survey.rm} must have been run to create a `pars.survey.rm' object, which contains the survey specifications.
}
\value{
\code{generate.sample.rm} returns an object of class `sample.rm´ which has the following elements:
  \item{population}{object of class `population´.}
  \item{design}{object of class `design.rm´.}
  \item{removal}{two dimensional detection matrix that indicates which animal groups have been captured on which occasion. A \code{1} in row i, column j indicates that the i-th animal/group was removed the occasion before j; a \code{0} means that it was not.}
  \item{detected}{two dimensional detection matrix that indicates which animal groups have been detected and consequently removed on which occasion. A \code{1} in row i, column j indicates that the i-th animal/group was detected on occasion j; a \code{0} means that it was not.}
  \item{parents}{Details of WiSP objects passed to function}
  \item{created}{Creation date and time}
  \item{seed}{Equal to the argument 'seed' passed to the function}
}
\seealso{
        \code{\link{setpars.survey.rm}},   \code{\link{summary.sample.rm}}
		\code{\link{plot.sample.rm}}
}
\examples{
rm.reg<-generate.region(x.length=100, y.width=50)
rm.dens <- generate.density(rm.reg)
rm.poppars<-setpars.population(density.pop = rm.dens, number.groups = 100,
                              size.method = "poisson", size.min = 1, size.max = 5,
                              size.mean = 1, exposure.method = "beta", exposure.min = 2,
                              exposure.max = 10, exposure.mean = 3, exposure.shape = 0.5,
                              type.values=c("Male","Female"), type.prob=c(0.48,0.52))
rm.pop<-generate.population(rm.poppars)

rm.des<-generate.design.rm(rm.reg, n.occ = 5, effort=c(1,2,3,2,1))
rm.survpars<-setpars.survey.rm(pop=rm.pop, des=rm.des, pmin=0.03, pmax=0.95, improvement=0.05)

rm.samp<-generate.sample.rm(rm.survpars)
plot(rm.samp)
plot(rm.samp,whole.population=TRUE)
}
\keyword{}

