\name{plot.design.no}
\alias{plot.design.no}
\title{Default Plotting Statement for Survey Design of Nearest-Object Method}
\description{
     This function generates a graphical representation of a point-to-nearest-object method design,  
generated by the function \code{generate.sample.no}.
}
\usage{
     plot(des, ...)
     plot.design.no(des, ...)
}
\arguments{
\item{des}{design object (of class `design.no´)}
}
\seealso{
     \code{\link{generate.design.no}}
}
\examples{
     myreg<-generate.region(x.length = 50, y.width = 80)
          
     mydens <- generate.density()
     mypoppars<-setpars.population(myreg, density.pop = mydens, 
                       number.groups = 100, size.method = "poisson", 
                       size.min = 1, size.max = 5, size.mean = 1,     
                       exposure.method = "beta", exposure.min = 2, 
                       exposure.max = 10, exposure.mean = 6,
                       exposure.shape = 1)
     mypop<-generate.population(mypoppars)
     mydes <- generate.design.no(myreg, n.points = 20)
     plot(mydes)
}
\keyword{}

