\name{point.est.no}
\alias{point.est.no}
\title{Point Estimation Of Population Size For Nearest-Object Method}
\description{
     This function calculates the maximum likelihood estimator of the abundance for the  
point-to-nearest-object method.
}
\usage{
    point.est.no(samp)
}
\arguments{
\item{samp}{object of class `sample.no´, generated by function 
	    \code{generate.sample.no}}
}
\details{
The point estimator of N is the MLE given by \eqn{\hat{N}=\frac{JA}{\pi\sum_j x_j^2}}{Nhat = J * A  
/ (pi * sum(xj^2))}
where
	
	J  = Number of survey points\cr
	A  = Total area of survey region\cr
	pi = 3.1415....\cr
	xj = The distances from survey point j to its nearest object\cr
}
\value{
An object of class `point.est.no´ containing the following items:

\item{Nhat}{MLE of group abundance}
}
\seealso{
     \code{\link{int.est.no}},
     \code{\link{generate.sample.no}} to create sample objects,
     \code{\link{summary.sample.no}},
     \code{\link{plot.sample.no}}      
}
\examples{
     
     # generating all necessary objects
     myreg<-generate.region(x.length = 50, y.width = 80)
          
     mydens <- generate.density()
     mypoppars<-setpars.population(myreg, density.pop = mydens, 
                          number.groups = 100, size.method = "poisson", 
                          size.min = 1, size.max = 5, size.mean = 1,     
                          exposure.method = "beta", exposure.min = 2, 
                          exposure.max = 10, exposure.mean = 6,
                          exposure.shape = 0.1)
     mypop<-generate.population(mypoppars)
     mydes <- generate.design.no(myreg, n.points = 4)
     mysamp<-generate.sample.no(mypop, mydes, with.neighbours=T)
     
     # calling estimation function
     point.est.no(mysamp)
     
}
\keyword{}

