\name{seal.pop}
\alias{seal.pop}
\title{A simulated seal population on a beach.}
\description{
These are simulated data are based on an aerial photograph of a real seal population on a beach.
The data are in a wisp object of class 'population' containing: 

\itemize{
 \item{region}{object of class `region'}
 \item{groupID}{vector of identification numbers for the animal groups.} 
 \item{posx}{vector of x-positions of the animal groups.}
 \item{posy}{vector of y-positions of the animal groups.}
 \item{groupsize}{vector of sizes of the animal groups.}
 \item{ntypes}{number of possible type properties for the animal groups.}
 \item{types}{vector of type properties of the animal groups.}
 \item{minexposure}{lowest possible group exposure value.}
 \item{maxexposure}{highest possible group exposure value.}
 \item{exposure}{vector of exposure values of the animal groups.}
 \item{parents}{Details of WiSP objects passed to function.}
 \item{created}{Creation date and time.}
 \item{seed}{Random number seed used to create the object.}
}}
\usage{data(seal.pop)}
\format{An object of class `population`}
\source{Aerial photograph}
\references{None}
\keyword{datasets}

