\name{setpars.design.dp}
\alias{setpars.design.dp}
\title{Survey Design of Double Platform Method }
\description{
        This function stores the information needed to define a survey design of the double platform method in a `pars.design.dp´ object without further modification.
}
\usage{
        setpars.design.dp(reg, n.transects = 1, n.units = 1, visual.range, percent.on.effort = 1)
}
\arguments{
\item{reg}{object of class `region´.}
\item{n.transects}{number of transect paths.}
\item{n.units}{number of survey units.}
\item{visual.range}{visibility of observer.}
\item{percent.on.effort}{percentage part of the transect paths that is actually used for observation.}
}
\details{
        A survey design of the double platform method is defined. The survey units lie along a transect path. This path need not be the entire survey unit, only part of it can be used for observation. The number of detected animals provides information on the size of the population. Detections are not assumed to be certain. The further the distance between transect path and an animal group the lower is the detection probability.
The transect paths run parallel to the y-axis and are distributed equidistant from each other, alternately changing their direction. Their common displacement on the x-axis is calculated randomly. The survey units have all the same dimension, are distributed equidistantly to each other and each transect path contains an equal number of them. The x-positions of the survey units are determined by the x-positions of the transect paths, the vertical displacement is calculated randomly.

        The `pars.design.dp´ object already contains all elementary survey design parameters as specified by the arguments of the function. However, an object of class `design.dp´ can only be generated by running the function \code{generate.design.dp}.
}
\value{
An object of class `pars.design.dp´ which has to be passed on to the function \code{generate.design.dp} and containing the following elements:
  \item{region}{the region over which the survey takes place}
  \item{n/transects}{the number of transects}
  \item{n.units}{the number of survey units}
  \item{visual.range}{visibility of the observer}
  \item{percentage.on.effort}{percentage of the transect paths that is actually used for observation}
  \item{parents}{details of the WiSP objects passed to the function}
  \item{created}{Creation date and time}
}
\seealso{
        \code{\link{generate.region}} to create regions
     	\code{\link{generate.design.dp}},   \code{\link{plot.design.dp}}
}
\examples{
dp.reg <- generate.region(x.length = 100, y.width = 50)
dp.dens <- generate.density(dp.reg)

#heterogeneous population
dp.poppars<-setpars.population(density.pop = dp.dens, number.groups = 1000, size.method = "poisson",	
						size.min = 1, size.max = 30, size.mean = 10, exposure.method = "beta", 
						exposure.min = 0, exposure.max = 1, exposure.mean = 0.4, exposure.shape = 0.5, 
						type.values=c("Male","Female"), type.prob=c(0.48,0.52))
dp.pop<-generate.population(dp.poppars)

dp.despars<-setpars.design.dp(dp.reg, n.transects=10, n.units=10, visual.range=2, percent.on.effort=1)
summary(dp.despars)
}
\keyword{}

