\name{setpars.survey.dp}
\alias{setpars.survey.dp}
\title{Survey Sample Parameters of Double Platform Method}
\description{
        This function stores the information needed to define a survey sample of the double platform method in a `pars.survey.dp´ object.
}
\usage{
        setpars.survey.dp(pop, des, adjust.interactive = TRUE, theta.obs1 = 1, theta.obs2 = 1, theta.exp = 0.1, theta.dist = -1, ...)
}
\arguments{
\item{pop}{object of class `population´.}
\item{des}{object of class `design.dp´.}
\item{adjust.interactive}{flag that enables interactive adjustment theta-parameters.}
\item{theta.obs1}{adjustment parameter for observer 1 in the detection function.}
\item{theta.obs2}{adjustment parameter for observer 2 in the detection function.}
\item{theta.exp}{adjustment parameter for exposure values in the detection function.}
\item{theta.dist}{adjustment parameter for animal groups distances in the detection function.}
\item{...}{extra plot arguments}
}
\details{
        This function calculates the detection probability of an animal group by applying a detection function. For the double platform method this function depends on the exposure of the animal groups, the distance of an animal group from the line and the visibility of the observers. This information is included in the population and design objects.  Furthermore, the detection function is specified by parameters that are directly given by the arguments `theta.obs1', 'theta.obs2', 'theta.exp' and 'theta.dist'. Compared to the other observation methods, the parameters are not longer calculated indirectly but have to be entered directly because the detection function is more sophisticated. The 'pars.survey.dp´ object already contains all elementary sample parameters as specified by the arguments of the function. However, an object of class `sample.dp´ can only be generated by running the function \code{generate.sample.dp}.
}
\value{ 
Returns an object of class `pars.survey.dp´ which has to be passed on to the function \code{generate.sample.dp} as a parameter.  The object is a list containing some or all of the following information: (population, design, theta.obs1, theta.obs2, theta.exp, theta.dist, parents, created)
}
\seealso{
        \code{\link{setpars.population}},   \code{\link{generate.population}} 
     	\code{\link{setpars.design.dp}}, 	\code{\link{generate.design.dp}} 
     	\code{\link{generate.sample.dp}},	\code{\link{summary.sample.dp}}, \code{\link{plot.sample.dp}}
}
\examples{
dp.reg <- generate.region(x.length = 100, y.width = 50)
dp.dens <- generate.density(dp.reg)

#heterogeneous population
dp.poppars<-setpars.population(density.pop = dp.dens, number.groups = 1000, size.method = "poisson", 
						size.min = 1, size.max = 30, size.mean = 10, exposure.method = "beta", 
						exposure.min = 0, exposure.max = 1, exposure.mean = 0.4, exposure.shape = 0.5, 
						type.values=c("Male","Female"), type.prob=c(0.48,0.52))
dp.pop<-generate.population(dp.poppars)
dp.despars<-setpars.design.dp(dp.reg, n.transects=10, n.units=10, visual.range=2, percent.on.effort=1) 
dp.des<-generate.design.dp(dp.despars, seed=3)

dp.survpars<-setpars.survey.dp(dp.pop, dp.des, theta.obs1=0.7, theta.obs2=0.4, theta.exp=0.2, theta.dist=-0.7)
summary(dp.survpars)
				 
# to set detection function parameters interactively:
# for a key press 'h' when inside
dp.survpars<-setpars.survey.dp(dp.pop, dp.des, adjust.interactive =TRUE, theta.obs1=0.7, theta.obs2=0.4, theta.exp=0.2, theta.dist=-0.7)
}
\keyword{}

