\name{skink.samp.cr} 
\alias{skink.samp.cr}
\title{Mark-recapture sample of spotted skinks on North Brother Island, New Zealand}
\description{The data are from a 7-sample mark-recapture survey of skinks \emph{(Oligosoma  
lineoocellatum)} on North Brother Island

\itemize{
\item{"population"}{Arbitrary population information that is required by the sample  
object - you can ignore it}
\item{"design"}{Arbitrary design information required by the sample object - you can  
ignore it}
\item{"capture"}{
A 168 * 7 matrix of capture histories of 1s and 0s where the rows represent the  
distinct skinks 
and the columns represent the capture occasions.  If skink i was caught on occasion  
j then the
(i,j)th entry of the capture matrix will be 1; otherwise it will be 0} 
\item{parents}{Information that is required by WiSP - you can ignore it.}
\item{created}{Information that is required by WiSP - you can ignore it.}
\item{seed}{Information that is required by WiSP - you can ignore it.}
}}
\usage{data(skink.samp.cr)} 
\format{An R object of class'sample.cr'} 
\source{P. Phillpot. 2000. unpublished M.Sc. thesis, Victoria University of  
Wellington. 
Skinks \emph{(Oligosoma lineoocellatum)} on North Brother Island, November 1999.}
\references{P. Phillpot. 2000. unpublished M.Sc. thesis, Victoria University of  
Wellington. 
Skinks \emph{(Oligosoma lineoocellatum)} on North Brother Island, November 1999.}
\keyword{datasets}

